Added new policy, updated controllers, small repo changes

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,17 +1,15 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
-using Restaurants.Domain;
+using Restaurants.Domain.Constants;
+using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurantCommand
 {
     public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-        IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+        IRestaurantsRepository restaurantsRepository,
+        IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
     {
         public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
         {
@@ -19,6 +17,9 @@ namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurantCommand
             var restaurant = await restaurantsRepository.GetById(request.Id );
             if ( restaurant is null ) 
                 throw new NotFoundException($"Restaurant", request.Id.ToString());
+
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+                throw new ForbidExcpetion();
 
             await restaurantsRepository.Delete(restaurant);
         }
